Move vector functions to vector.c
