+++ src/stego.h
@@ -16,6 +16,4 @@ int stego_read_level(const void* key, bs_disk_t disk, void* buf, off_t off,
 int stego_write_level(const void* key, bs_disk_t disk, const void* buf,
                       off_t off, size_t size);
 
-int stego_gen_keys(void* buf, int count);
-
 #endif // BS_STEGO_H
